Squashed version of/replaces #9075

## Description

Implementation overview: https://gist.github.com/klauspost/1801c858d5e0df391114436fdad6987b

Includes staticcheck upgrade and fixes for that.

Some quick performance tests of crawls.

This is with a cycle size of 16, 174082 XL objects. ~4K folders at prefix level 2, NVME.

```
BEFORE:
Crawl time 1m17.3025259s
Disk access: 350MB/s
Kernel  Time =    59.468 =   70%
User    Time =    21.984 =   25%
Process Time =    81.453 =   95%    Virtual  Memory =    303 MB

AFTER:
SET MINIO_DISK_USAGE_CRAWL_DELAY=10 (default)
Cycle scan time: 45.354097s
Disk Access: 15MB/s
Kernel  Time =     0.562 =    1%
User    Time =     0.640 =    1%
Process Time =     1.203 =    2%    Virtual  Memory =    304 MB

SET MINIO_DISK_USAGE_CRAWL_DELAY=1
Cycle scan time: 3.3367481s
Disk Access: 160MB/s
Kernel  Time =     0.468 =    3%
User    Time =     1.421 =   12%
Process Time =     1.890 =   16%    Virtual  Memory =    303 MB

SET MINIO_DISK_USAGE_CRAWL_DELAY=0
Cycle scan time: 2.1245395s
Disk Access: 175MB/s
Kernel  Time =     1.500 =   14%
User    Time =     0.765 =    7%
Process Time =     2.265 =   22%    Virtual  Memory =    304 MB
```

## How to test this PR?

For now the server will display extra information on crawling.

## Types of changes
- [x] New feature (non-breaking change which adds functionality)